+++ apps/test-can/src/main.cpp
@@ -9,8 +9,8 @@ DigitalOut ledRX(LED2);
 Timer timer;
 uint8_t counter = 0;
 
-constexpr int TX_ID = 0x100;
-constexpr int RX_ID = 0x101;
+constexpr uint16_t TX_ID = 0x100;
+constexpr uint16_t RX_ID = 0x101;
 
 /**
  * @brief   Prints CAN msg to PC's serial terminal
